chore(breakdown-tables): break into ethnicity-table.js and race-table.js
